+++ src/test/java/pe/dcalma/springboot/todolist/UsuarioServiceTest.java
@@ -1,0 +1,122 @@
+package pe.dcalma.springboot.todolist;
+
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import pe.dcalma.springboot.todolist.model.Usuario;
+import pe.dcalma.springboot.todolist.service.UsuarioService;
+import pe.dcalma.springboot.todolist.service.UsuarioServiceException;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+@ExtendWith(SpringExtension.class)
+@SpringBootTest
+public class UsuarioServiceTest {
+
+    Logger logger = LoggerFactory.getLogger(UsuarioServiceTest.class);
+
+    @Autowired
+    private UsuarioService usuarioService;
+
+    @Test
+    public void servicioLoginUsuario() {
+        // GIVEN
+        // Datos cargados de datos-test.sql
+
+        // WHEN
+
+        UsuarioService.LoginStatus loginStatusOK = usuarioService.login("dcalma@gmail.com", "12345678");
+        UsuarioService.LoginStatus loginStatusErrorPassword = usuarioService.login("dcalma@gmail.com", "000");
+        UsuarioService.LoginStatus loginStatusNoUsuario = usuarioService.login("pepito.perez@gmail.com", "12345678");
+
+        // THEN
+
+        assertThat(loginStatusOK).isEqualTo(UsuarioService.LoginStatus.LOGIN_OK);
+        assertThat(loginStatusErrorPassword).isEqualTo(UsuarioService.LoginStatus.ERROR_PASSWORD);
+        assertThat(loginStatusNoUsuario).isEqualTo(UsuarioService.LoginStatus.USER_NOT_FOUND);
+    }
+
+    @Test
+    public void servicioRegistroUsuario() {
+        // GIVEN
+
+        Usuario usuario = new Usuario("usuario.prueba2@gmail.com");
+        usuario.setPassword("12345678");
+
+        // WHEN
+
+        usuarioService.registrar(usuario);
+
+        // THEN
+
+        Usuario usuarioBaseDatos = usuarioService.findByEmail("usuario.prueba2@gmail.com");
+        assertThat(usuarioBaseDatos).isNotNull();
+        assertThat(usuarioBaseDatos.getPassword()).isEqualTo(usuario.getPassword());
+    }
+
+    @Test
+    public void servicioRegistroUsuarioExcepcionConNullPassword() {
+        // Pasamos como argumento un usario sin contraseña
+        Usuario usuario =  new Usuario("usuario.prueba@gmail.com");
+
+        assertThrows(UsuarioServiceException.class, () -> {
+            usuarioService.registrar(usuario);
+        });
+
+    }
+
+
+    @Test
+    public void servicioRegistroUsuarioExcepcionConEmailRepetido() {
+        // GIVEN
+        // Datos cargados de datos-test.sql
+
+        // WHEN
+        // Pasamos como argumento un usario con emaii existente en datos-test.sql
+        Usuario usuario =  new Usuario("dcalma@gmail.com");
+        usuario.setPassword("12345678");
+
+        assertThrows(UsuarioServiceException.class, () -> {
+            usuarioService.registrar(usuario);
+        });
+
+        // THEN
+        // Se produce una excepción comprobada con el expected del test
+    }
+
+    @Test
+    public void servicioRegistroUsuarioDevuelveUsuarioConId() {
+        // GIVEN
+
+        Usuario usuario = new Usuario("usuario.prueba@gmail.com");
+        usuario.setPassword("12345678");
+
+        // WHEN
+
+        usuario = usuarioService.registrar(usuario);
+
+        // THEN
+
+        assertThat(usuario.getId()).isNotNull();
+    }
+
+    @Test
+    public void servicioConsultaUsuarioDevuelveUsuario() {
+        // GIVEN
+        // Datos cargados de datos-test.sql
+
+        // WHEN
+
+        Usuario usuario = usuarioService.findByEmail("dcalma@gmail.com");
+
+        // THEN
+
+        assertThat(usuario.getId()).isEqualTo(1L);
+    }
+}
